code/petCar/PythonCode/requirements.txt
python-dotenv
huggingface_hub
pillow
openai
requests
flask
opencv-python